+++ src/query/formats/src/field_decoder/fast_values.rs
@@ -65,6 +65,7 @@ use lexical_core::FromLexical;
 use num::cast::AsPrimitive;
 use num_traits::NumCast;
 
+use crate::field_decoder::common::read_timestamp;
 use crate::FieldDecoder;
 use crate::InputCommonSettings;
 
@@ -323,26 +324,7 @@ impl FastFieldDecoderValues {
     ) -> Result<()> {
         let mut buf = Vec::new();
         self.read_string_inner(reader, &mut buf, positions)?;
-        let mut buffer_readr = Cursor::new(&buf);
-        let ts = buffer_readr.read_timestamp_text(&self.common_settings().jiff_timezone)?;
-        match ts {
-            DateTimeResType::Datetime(ts) => {
-                if !buffer_readr.eof() {
-                    let data = buf.to_str().unwrap_or("not utf8");
-                    let msg = format!(
-                        "fail to deserialize timestamp, unexpected end at pos {} of {}",
-                        buffer_readr.position(),
-                        data
-                    );
-                    return Err(ErrorCode::BadBytes(msg));
-                }
-                let mut micros = ts.timestamp().as_microsecond();
-                clamp_timestamp(&mut micros);
-                column.push(micros.as_());
-            }
-            _ => unreachable!(),
-        }
-        Ok(())
+        read_timestamp(column, &buf, &self.common_settings())
     }
 
     fn read_array<R: AsRef<[u8]>>(
