+++ src/query/formats/src/field_decoder/mod.rs
@@ -12,6 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+mod common;
 mod fast_values;
 mod json_ast;
 mod nested;
